create models

+++ database/migrations/2025_01_21_164803_category.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('category', function (Blueprint $table) {
             $table->id();
-            $table->integer('parent_id')->nullable();
+            $table->unsignedBigInteger('parent_id')->nullable();
             $table->string('key');
             $table->string('title');
             $table->string('slug');
